requirements.txt
ccxt
pandas
numpy
matplotlib
streamlit
python-dotenv
streamlit-option-menu 
streamlit-lottie 
requests
